requirements.txt
python-telegram-bot>=22.5
requests>=2.32.0



